+++ src/org/xjge/core/DebugInfo.java
@@ -1,6 +1,5 @@
 package org.xjge.core;
 
-import org.xjge.graphics.Texture;
 import org.xjge.graphics.Rectangle;
 import org.xjge.graphics.Icon;
 import org.xjge.graphics.RectangleBatch;
@@ -14,6 +13,8 @@ import static org.lwjgl.openal.AL10.alGetString;
 import static org.lwjgl.opengl.GL11.GL_RENDERER;
 import static org.lwjgl.opengl.GL11.GL_VERSION;
 import static org.lwjgl.opengl.GL11.glGetString;
+import static org.xjge.core.Font.defaultFont;
+import org.xjge.graphics.Texture;
 
 /**
  * Created: Jun 10, 2021
@@ -30,9 +31,6 @@ final class DebugInfo {
     
     boolean show;
     
-    private final Font font;
-    private final Text text = new Text();
-    
     private final RectangleBatch rectBatch = new RectangleBatch(6);
     
     private final Rectangle[] rectangles = new Rectangle[6];
@@ -45,9 +43,7 @@ final class DebugInfo {
      * @param font the font supplied by the engine that the interface will use
      * @param iconTexture the texture containing the various icons in the interface
      */
-    DebugInfo(Font font, Texture iconTexture) {
-        this.font = font;
-        
+    DebugInfo(Texture iconTexture) {
         for(int r = 0; r < rectangles.length; r++) rectangles[r] = new Rectangle();
         for(int p = 0; p < textPos.length; p++)    textPos[p] = new Vector2i();
         for(int i = 0; i < icons.length; i++)      icons[i] = new Icon(iconTexture, 64, 64);
@@ -84,7 +80,7 @@ final class DebugInfo {
                 if(string.length() > longestString.length()) longestString = string;
             }
             
-            rectangles[1].width  = Text.lengthInPixels(longestString, font) + (PADDING * 2);
+            rectangles[1].width  = defaultFont.lengthInPixels(longestString) + (PADDING * 2);
             rectangles[1].height = (Font.DEFAULT_SIZE * 3) + PADDING;
             rectangles[1].xPos   = Window.getWidth() - rectangles[1].width;
             rectangles[1].yPos   = Window.getHeight() - rectangles[1].height;
@@ -108,7 +104,7 @@ final class DebugInfo {
                 if(string.length() > longestString.length()) longestString = string;
             }
             
-            rectangles[2].width  = Text.lengthInPixels(longestString, font) + (PADDING * 2);
+            rectangles[2].width  = defaultFont.lengthInPixels(longestString) + (PADDING * 2);
             rectangles[2].height = (Font.DEFAULT_SIZE * 4) + PADDING;
             rectangles[2].xPos   = Window.getWidth() - rectangles[2].width;
             rectangles[2].yPos   = rectangles[1].yPos - (rectangles[2].height + PADDING);
@@ -133,7 +129,7 @@ final class DebugInfo {
                 if(string.length() > longestString.length()) longestString = string;
             }
             
-            rectangles[3].width  = Text.lengthInPixels(longestString, font) + (PADDING * 2);
+            rectangles[3].width  = defaultFont.lengthInPixels(longestString) + (PADDING * 2);
             rectangles[3].height = (Font.DEFAULT_SIZE * 4) + PADDING;
             rectangles[3].xPos   = Window.getWidth() - rectangles[3].width;
             rectangles[3].yPos   = rectangles[2].yPos - (rectangles[3].height + PADDING);
@@ -156,7 +152,7 @@ final class DebugInfo {
                 }
             }
             
-            int length = Text.lengthInPixels(longestString, font) + (64 + (PADDING * 2));
+            int length = defaultFont.lengthInPixels(longestString) + (64 + (PADDING * 2));
             
             rectangles[4].width  = (length > rectangles[0].width) ? length : rectangles[0].width;
             rectangles[4].height = 252;
@@ -181,7 +177,7 @@ final class DebugInfo {
         
         //Group 6: Engine Info
         {
-            rectangles[5].width  = Text.lengthInPixels("XJGE v" + XJGE.VERSION, font) + (PADDING * 2);
+            rectangles[5].width  = defaultFont.lengthInPixels("XJGE v" + XJGE.VERSION) + (PADDING * 2);
             rectangles[5].height = 24 + PADDING;
             rectangles[5].xPos   = Window.getWidth() - rectangles[5].width;
             rectangles[5].yPos   = rectangles[3].yPos - (rectangles[5].height + PADDING);
@@ -211,37 +207,38 @@ final class DebugInfo {
             icons[i].render();
         }
         
-        text.drawString(font, "FPS: " + Game.getFPS(), textPos[0], Color.WHITE, 1f);
-        text.drawString(font, "DELTA: " + Game.getDelta(), textPos[1], Color.WHITE, 1f);
-        text.drawString(font, "TICKED: " + Game.getTicked(), textPos[2], Color.WHITE, 1f);
-        text.drawString(font, "MEM FREE: " + Runtime.getRuntime().freeMemory(), textPos[3], Color.CYAN, 1f);
+        defaultFont.drawString("FPS: " + Game.getFPS(), textPos[0].x, textPos[0].y, Color.WHITE, 1f);
+        defaultFont.drawString("DELTA: " + Game.getDelta(), textPos[1].x, textPos[1].y, Color.WHITE, 1f);
+        defaultFont.drawString("TICKED: " + Game.getTicked(), textPos[2].x, textPos[2].y, Color.WHITE, 1f);
+        defaultFont.drawString("MEM FREE: " + Runtime.getRuntime().freeMemory(), textPos[3].x, textPos[3].y, Color.CYAN, 1f);
         
-        text.drawString(font, "OS NAME: " + System.getProperty("os.name"), textPos[4],  Color.GREEN, 1f);
-        text.drawString(font, "JAVA VER: " + System.getProperty("java.version"), textPos[5],  Color.GREEN, 1f);
-        text.drawString(font, "GLFW VER: " + glfwGetVersionString(), textPos[6], Color.GREEN, 1f);
+        defaultFont.drawString("OS NAME: " + System.getProperty("os.name"), textPos[4].x, textPos[4].y,  Color.GREEN, 1f);
+        defaultFont.drawString("JAVA VER: " + System.getProperty("java.version"), textPos[5].x, textPos[5].y,  Color.GREEN, 1f);
+        defaultFont.drawString("GLFW VER: " + glfwGetVersionString(), textPos[6].x, textPos[6].y, Color.GREEN, 1f);
         
-        text.drawString(font, "OPENAL VER: " + alGetString(AL_VERSION), textPos[7], Color.WHITE, 1f);
-        text.drawString(font, "SPEAKER: \"" + Audio.speaker.name.substring(15) + "\"", textPos[8], Color.YELLOW, 1f);
-        text.drawString(font, "SOUND MASTER VOLUME: " + Audio.getSoundMasterVolume(), textPos[9], Color.CYAN, 1f);
-        text.drawString(font, "MUSIC MASTER VOLUME: " + Audio.getMusicMasterVolume(), textPos[10], Color.CYAN, 1f);
+        defaultFont.drawString("OPENAL VER: " + alGetString(AL_VERSION), textPos[7].x, textPos[6].y, Color.WHITE, 1f);
+        defaultFont.drawString("SPEAKER: \"" + Audio.speaker.name.substring(15) + "\"", textPos[8].x, textPos[8].y, Color.YELLOW, 1f);
+        defaultFont.drawString("SOUND MASTER VOLUME: " + Audio.getSoundMasterVolume(), textPos[9].x, textPos[9].y, Color.CYAN, 1f);
+        defaultFont.drawString("MUSIC MASTER VOLUME: " + Audio.getMusicMasterVolume(), textPos[10].x, textPos[10].y, Color.CYAN, 1f);
         
-        text.drawString(font, "OPENGL VER: " + glGetString(GL_VERSION), textPos[11], Color.WHITE, 1f);
-        text.drawString(font, "GFX CARD: " + glGetString(GL_RENDERER), textPos[12], Color.YELLOW, 1f);
-        text.drawString(font, "DISPLAY: \"" + Window.monitor.name + "\" (" + Window.monitor.getInfo() + ")", textPos[13], Color.YELLOW, 1f);
-        text.drawString(font, "VSYNC: " + Hardware.getVSyncEnabled(), textPos[14], Color.CYAN, 1f);
+        defaultFont.drawString("OPENGL VER: " + glGetString(GL_VERSION), textPos[11].x, textPos[11].y, Color.WHITE, 1f);
+        defaultFont.drawString("GFX CARD: " + glGetString(GL_RENDERER), textPos[12].x, textPos[11].y, Color.YELLOW, 1f);
+        defaultFont.drawString("DISPLAY: \"" + Window.monitor.name + "\" (" + Window.monitor.getInfo() + ")", 
+                               textPos[13].x, textPos[13].y, Color.YELLOW, 1f);
+        defaultFont.drawString("VSYNC: " + Hardware.getVSyncEnabled(), textPos[14].x, textPos[14].y, Color.CYAN, 1f);
         
-        text.drawString(font, Input.getDeviceEnabled(KEY_MOUSE_COMBO) ? "Enabled" : "Disabled", textPos[15], Color.WHITE, 1f);
+        defaultFont.drawString(Input.getDeviceEnabled(KEY_MOUSE_COMBO) ? "Enabled" : "Disabled", 
+                               textPos[15].x, textPos[15].y, Color.WHITE, 1f);
         
         for(int i = 16; i < textPos.length - 1; i += 2) {
             int jid = (i % 16) / 2;
             
-            text.drawString(font, jid + "", textPos[i], Color.YELLOW, 1f);
-            text.drawString(font, Input.getDevicePresent(jid) ? Input.getDeviceName(jid) : "N/A", textPos[i + 1], Color.WHITE, 1f);
+            defaultFont.drawString(jid + "", textPos[i].x, textPos[i].y, Color.YELLOW, 1f);
+            defaultFont.drawString(Input.getDevicePresent(jid) ? Input.getDeviceName(jid) : "N/A", 
+                                   textPos[i + 1].x, textPos[i + 1].y, Color.WHITE, 1f);
         }
         
-        text.drawString(font, "XJGE v" + XJGE.VERSION, textPos[24], Color.MAGENTA, 1f);
-        
-        text.resetStringIndex();
+        defaultFont.drawString("XJGE v" + XJGE.VERSION, textPos[24].x, textPos[14].y, Color.MAGENTA, 1f);
     }
 
     /**